Added known_structures (archive and archive II only).
Dave:
--------
archiveII is what you should be using now.  It is the same families we have had in archive, but updated with the latest databases.  Mike Sloma, a graduate student, did the work.

I am pretty sure that all the filenames now start with the RNA family name and end with the database accession #.  The latest reference that details the families is:
Bellaousov, S., & Mathews, D. H. (2010). ProbKnot: fast prediction of RNA secondary structure including pseudoknots. RNA. 16, 1870-1880.

Another database you should also fetch from cvs is: RNA_stralign.  That database has structures and alignments for eight families, and a ReadMe that provides a lot of detail.
---------